+++ includes_test/includes_test_win/main.cpp
@@ -6,9 +6,11 @@ using namespace std;
 
 int main()
 {
+	ControlInit();
+
 	while(true)
 	{
-		cout << ControlStates() << endl;
+		cout << ControlStates(true) << endl;
 		Sleep(1000/60);
 	}
     return 0;