+++ native_library/native_library_mac/AIRControl/AIRControl/AIRControl.cpp
@@ -38,9 +38,16 @@ extern "C"
 	FREObject AIRControl(FREContext ctx, void * funcData, uint32_t argc, FREObject argv[])
 	{
 		FREObject result;
-		std::string str(ControlStates());
-		const uint8_t * msg = (const uint8_t *)str.c_str();
-		FRENewObjectFromUTF8(strlen((const char *)msg)+1, msg, &result);
+		uint32_t debugModeFRE = 0;
+		bool debugMode = false;
+		//If the optional debug argument is defined, check if in debug mode.
+		if(argc)
+		{
+			FREGetObjectAsBool(argv[0], &debugModeFRE);
+			debugMode = debugModeFRE;
+		}
+		const char * str = ControlStates(debugMode).c_str();
+		FRENewObjectFromUTF8(strlen(str)+1, (const uint8_t *)str, &result);
 		
 		return result;
 	}