+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+1.1.0
+======
+
+* Limited the Windows extension to checking for new controllers to only once per second to improve performance.
+* Added debug output that can be enabled for runtime debugging.
+
+
+
+
 1.0.0
 ======